Added test for selecting a product.  If the total value of inserted coins is below the selected price of the product, display "INSERT COIN" Added Product object to contain information related to products. Added items to VendingMachine -Products, list of products in the vending machine -Display string to Vending Machine. -InsertedCoinsValue to hold the value of inserted coins. Added method SelectProduct which accepts the button being pressed.

+++ src/com/wolfhaus/Coin.java
@@ -3,9 +3,9 @@ package com.wolfhaus;
 public class Coin {
     protected int Size;
     protected int Weight;
-    protected int Value;
+    protected double Value;
 
-    public Coin(int coinSize, int coinWeight, int coinValue)
+    public Coin(int coinSize, int coinWeight, double coinValue)
     {
         this.Size = coinSize;
         this.Weight = coinWeight;
